requirements.txt
colorama==0.4.4
Cython==0.29.24
easydict==1.9
hyperopt==0.2.7
kornia==0.6.4
lmdb==1.2.1
matplotlib==3.5.0
mmdet==2.25.1
mpi4py==3.1.3
numpy==1.22.4
pandas==1.3.4
Pillow==9.2.0
prefetch_generator==1.0.1
pycocotools==2.0+nv0.6.0
PySide2==5.15.2.1
PyYAML==6.0
ray==2.0.0
scipy==1.6.3
setuptools==59.4.0
Shapely==1.8.1.post1
six
thop
tqdm
glob
opencv-python
numba
